Backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
Django==4.0.3
djangorestframework==3.13.1
djongo==1.3.6
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.22.2
pandas==1.4.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pymongo==4.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.1.1
